final sg90 code

+++ Code/Motor Driver/spsp/sg90.h
@@ -20,7 +20,7 @@ typedef enum sg90_state {
 
 void sg90_init_pins(void); //setup the pins and timers
 
-void sg90_init_timers(uint8_t angle); //angle 0-64
+void sg90_init_timers(uint16_t angle); //angle 0-64
 
 void sg90_move(const enum sg90_state state);
